feat: added material parameter
